Split render material and texture caches
