+++ src/render/wgpu/mod.rs
@@ -11,12 +11,14 @@ pub mod ltc;
 pub mod material;
 pub mod matrix;
 pub mod mesh;
+pub mod render_material_cache;
 pub mod render_gizmo_item;
 pub mod render_item;
 pub mod render_light_item;
 pub mod render_mesh_item;
 pub mod render_object_instance_item;
 pub mod render_resource;
+pub mod render_texture_cache;
 pub mod shader;
 pub mod shadow;
 pub mod shadow_map_renderer;
